examples/mcp/streamable-http/ai-foundry-agent/pyproject.toml
[project]
name = "ai-foundry-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-agents>=1.0.1",
    "azure-ai-projects>=1.0.0b12",
    "azure-identity>=1.23.0",
    "python-dotenv>=1.1.1",
]
